2819 push new avatar url untill middleware
