+++ android/sdk/src/main/java/org/jitsi/meet/sdk/JitsiMeetView.java
@@ -256,7 +256,6 @@ public class JitsiMeetView extends BaseReactView<JitsiMeetViewListener>
      * Update avatar url in thumbnails.
      */
     public void updateAvatar(String dataJsonString) {
-        JitsiMeetLogger.w(dataJsonString + " ------------------- IMAGE " );
         VideoConfBridge.updateAvatar(dataJsonString);
     }
